% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_db.R
\name{acep_db}
\alias{acep_db}
\title{Frecuencia, menciones e intensidad}
\usage{
acep_db(db, t, d, n)
}
\arguments{
\item{db}{data frame con los textos a procesar.}

\item{t}{columna de data frame que contiene el vector de textos a procesar.}

\item{d}{diccionario en formato vector.}

\item{n}{cantidad de decimales del indice de intensidad.}
}
\description{
Funcion que usa las funciones acep_frec, acep_men y acep_int y devuelve una tabla con tres columnas nuevas: numero de palabras, numero de menciones del diccionario, indice de intensidad.
}
\examples{
rp_procesada <- acep_db(rev_puerto, rev_puerto$nota, diccionarios$dicc_viol_gp, 4)
}
