% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_extract.R
\name{acep_extract}
\alias{acep_extract}
\title{Funcion para buscar y extraer palabras en un texto}
\usage{
acep_extract(texto, dicc, sep = " | ", tolower = TRUE)
}
\arguments{
\item{texto}{El texto en el que se buscarán las palabras clave.}

\item{dicc}{Un vector con las palabras clave a buscar.}

\item{sep}{El separador utilizado para concatenar las
palabras clave encontradas (por defecto: " | ").}

\item{tolower}{convierte los textos a minusculas.}
}
\description{
Esta funcion busca palabras clave en un texto y
extrae los resultados en un formato especifico.
}
\examples{
texto <- "Los obreros del pescado, en el marco de una huelga y
realizaron una manifestación con piquete en el puerto de la ciudad."
dicc <- c("huel", "manif", "piq")
acep_extract(texto, dicc)
}
\keyword{buscar}
\keyword{diccionario}
\keyword{palabras}
\keyword{texto}
