% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_frec.R
\name{acep_frec}
\alias{acep_frec}
\title{Frecuencia de menciones de palabras que refieren a conflictos.}
\usage{
acep_frec(x)
}
\arguments{
\item{x}{vector de textos al que se le aplica la función de conteo de la frecuencia de menciones de palabras que refieren a conflictos.}
}
\description{
Función que cuenta la frecuencia de menciones de palabras que refieren a conflictos en cada una de las notas/textos.
}
\examples{
notas$conflictos <- acep_frec(notas$nota)

}
