% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_int.R
\name{acep_int}
\alias{acep_int}
\title{Indice de intensidad.}
\usage{
acep_int(pc, pt, decimales = 4)
}
\arguments{
\item{pc}{vector numerico con la frecuencia de palabras conflictivas presentes en cada texto.}

\item{pt}{vector de palabras totales en cada texto.}

\item{decimales}{cantidad de decimales, por defecto tiene 4 pero se puede modificar.}
}
\description{
Funcion que elabora un indice de intensidad en base a la relacion entre palabras totales y palabras del diccionario presentes en el texto.
}
\examples{
rev_puerto$intensidad <- acep_int(rev_puerto$conflictos, rev_puerto$n_palabras, 2)
}
