% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_may.R
\name{acep_may}
\alias{acep_may}
\title{Convierte caracteres a mayusculas}
\usage{
acep_may(x)
}
\arguments{
\item{x}{es un vector de texto (caracteres) que se desea convertir a mayúsculas.}
}
\value{
Devuelve un nuevo vector con todas las letras en mayúsculas.
}
\description{
Esta función toma un vector de texto y convierte todas las letras
minúsculas en mayúsculas, manteniendo el resto de los caracteres sin cambios.
}
\examples{
vector_texto <- c("soip", "cGt", "Sutna")
acep_may(vector_texto)
vector_numeros <- c(1, 2, 3, 4, 5)
acep_may(vector_numeros)
vector_mezclado <- c("sutna", 123, "Ate")
acep_may(vector_mezclado)
}
\keyword{caracteres}
\keyword{texto}
