% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_min.R
\name{acep_min}
\alias{acep_min}
\title{Convierte caracteres a minúsculas.}
\usage{
acep_min(x)
}
\arguments{
\item{x}{Un vector de texto (caracteres) que se desea convertir a minúsculas.}
}
\value{
Devuelve un nuevo vector con todas las letras en minúsculas.
}
\description{
Esta función toma un vector de texto y convierte todas las letras
mayusculas en minúsculas, manteniendo el resto de los caracteres sin cambios.
}
\examples{
vector_texto <- c("SUTEBA", "Sindicato", "PEN")
acep_min(vector_texto)
vector_numeros <- c(1, 2, 3, 4, 5)
acep_min(vector_numeros)
vector_mezclado <- c("Soip", 123, "CGT")
acep_min(vector_mezclado)
}
\keyword{caracteres}
\keyword{texto}
