% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_ollama.R
\name{acep_ollama_setup}
\alias{acep_ollama_setup}
\title{Guia de instalacion y uso de Ollama}
\usage{
acep_ollama_setup()
}
\description{
Imprime instrucciones para instalar y configurar Ollama en tu sistema.
}
