% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_plot_rst.R
\name{acep_plot_rst}
\alias{acep_plot_rst}
\title{Resumen visual de la serie temporal de los índices de conflictividad.}
\usage{
acep_plot_rst(db, tagx = "horizontal")
}
\arguments{
\item{db}{data frame con datos procesados.}

\item{tagx}{orientación de las etiquetas del eje x ('horizontal' | 'vertical').}
}
\description{
Función que devuelve un panel visual de cuatro gráficos de barras con variables proxy de los índices de conflictividad agrupados por segmento de tiempo.
}
\examples{
datos <- acep_db(acep_bases$rev_puerto, acep_bases$rev_puerto$nota, acep_diccionarios$dicc_viol_gp, 4)
datos_procesados_anio <- acep_rst(datos, datos$fecha, datos$n_palabras, datos$conflictos, st = 'anio')
acep_plot_rst(datos_procesados_anio, tagx = 'vertical')
}
