% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_sst.R
\name{acep_sst}
\alias{acep_sst}
\title{Serie temporal de índices de conflictividad.}
\usage{
acep_sst(datos, st = "mes", u = 2, d = 4)
}
\arguments{
\item{datos}{data frame con las variables 'fecha' (en formato Date),
'n_palabras' (numérica), conflictos' (numérica), 'intensidad' (numérica).
Las ultimas tres se pueden construir en un solo paso con la funcion 'acep_db'
o en tres pasos con las funciones 'acep_frec', 'acep_men', 'acep_int'.}

\item{st}{parámetro para establecer el segmento temporal
a ser agrupado: 'anio', 'mes', 'dia'.}

\item{u}{umbral de menciones para contabilizar una nota
como nota que refiere a un conflicto, por defecto tiene 2
pero se puede modificar.}

\item{d}{cantidad de decimales, por defecto tiene 4 pero
se puede modificar.}
}
\value{
Si todas las entradas son correctas,
la salida sera una base de datos en formato tabular
con nuevas variables.
}
\description{
Función que devuelve los indices de conflictividad
agrupados por segmento de tiempo: 'dia', 'mes', 'anio'. Esta función
viene a reemplazar a acep_rst. Simplifica los parámetros.
}
\examples{
datos <- acep_bases$rp_procesada
head(datos)
datos_procesados_anio <- acep_sst(datos, st='anio', u=4)
datos_procesados_mes <- acep_sst(datos)
datos_procesados_dia <- acep_sst(datos, st ='dia', d=3)
head(datos_procesados_anio)
head(datos_procesados_mes)
head(datos_procesados_dia)
}
\keyword{resumen}
