% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_token.R
\name{acep_token}
\alias{acep_token}
\title{Tokenizador.}
\usage{
acep_token(x, tolower = TRUE, cleaning = TRUE)
}
\arguments{
\item{x}{vector de textos al que se le aplica la función de tokenización.}

\item{tolower}{convierte los textos a minúsculas.}

\item{cleaning}{hace una limpieza de los textos.}
}
\value{
Si todas las entradas son correctas,
la salida será un data.frame con las palabras tokenizadas.
}
\description{
Función que tokeniza las notas/textos.
}
\examples{
acep_token("Huelga de obreros del pescado en el puerto")
}
\keyword{tokenizar}
