% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_token_table.R
\name{acep_token_table}
\alias{acep_token_table}
\title{Tabla de frecuencia de palabras tokenizadas.}
\usage{
acep_token_table(x, u = 10)
}
\arguments{
\item{x}{vector de palabras tokenizadas.}

\item{u}{número de corte para el top de palabras más frecuentes.}
}
\value{
Si todas las entradas son correctas,
la salida sera una tabla con la frecuencia relativa y
absoluta de palabras tokenizadas.
}
\description{
Función que cuenta la frecuencia de palabras tokenizadas.
}
\examples{
tokens <- c(rep("paro",15), rep("piquete",25), rep("corte",20), rep("manifestación",10),
rep("bloqueo",5), rep("alerta",16), rep("ciudad",12), rep("sindicato",11), rep("paritaria",14),
rep("huelga",14), rep("escrache",15))
acep_token_table(tokens)
}
\keyword{tablas}
\keyword{tokens}
