% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dias.R
\docType{data}
\name{dias}
\alias{dias}
\title{Filtro para remover días}
\format{
Es una cadena de caracteres.
\describe{
\item{dias}{es un string}
}
}
\usage{
data(dias)
}
\description{
Lista de palabras para remover días.
}
\examples{
print(dias)
}
\keyword{expresiones}
\keyword{regulares}
