% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionarios.R
\docType{data}
\name{dicc_conflictos}
\alias{dicc_conflictos}
\title{Colección de diccionarios.}
\format{
An object of class \code{character} of length 1325.
}
\usage{
dicc_conflictos
}
\description{
Función que carga un set de diccionarios de diferentes orígenes.
El diccionario de palabras que refieren a conflictos llamado dicc_conflictos
fue extraído de Palazzo (2017).
}
\examples{
diccionarios



}
\keyword{datasets}
