% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionarios.R
\docType{data}
\name{diccionarios}
\alias{diccionarios}
\title{Colección de diccionarios.}
\format{
An object of class \code{list} of length 2.
}
\usage{
diccionarios
}
\description{
Colección de diccionarios que reune diccionarios de diferentes orígenes. Los diccionarios dicc_confl_gp y dicc_viol_gp fueron extraídos de Albrieu y Palazzo (2020).
}
\examples{
dicc_confl_gp <- diccionarios$dicc_confl_gp
dicc_viol_gp <- diccionarios$dicc_viol_gp
}
\references{
Albrieu, Ramiro, y Gabriel Palazzo. 2020. «Categorización de conflictos sociales en el ámbito de los recursos naturales: un estudio de las actividades extractivas mediante la minería de textos». Revista CEPAL (131):29-59.
}
\keyword{datasets}
