% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_count.R
\name{.acep_count_cache_key}
\alias{.acep_count_cache_key}
\title{Crear una clave de caché estable para diccionarios de conteo}
\usage{
.acep_count_cache_key(dic)
}
\description{
Crear una clave de caché estable para diccionarios de conteo
}
\keyword{internal}
