% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_count.R
\name{.acep_escape_regex}
\alias{.acep_escape_regex}
\title{Escapar metacaracteres de expresiones regulares en términos literales}
\usage{
.acep_escape_regex(x)
}
\description{
Escapar metacaracteres de expresiones regulares en términos literales
}
\keyword{internal}
