% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_count.R
\docType{data}
\name{.acep_regex_cache}
\alias{.acep_regex_cache}
\title{Cache para regex compilados}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.acep_regex_cache
}
\description{
Cache para regex compilados
}
\keyword{internal}
