% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji.R
\docType{data}
\name{emoji}
\alias{emoji}
\title{Filtro para remover emojis}
\format{
Es una cadena de caracteres.
\describe{
\item{emoji}{es un string}
}
}
\usage{
data(emoji)
}
\description{
Expresiones regulares para remover emojis.
}
\examples{
print(emoji)
}
\keyword{expresiones}
\keyword{regulares}
