% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meses.R
\docType{data}
\name{meses}
\alias{meses}
\title{Filtro para remover meses}
\format{
Es una cadena de caracteres.
\describe{
\item{meses}{es un string}
}
}
\usage{
data(meses)
}
\description{
Lista de palabras para remover meses.
}
\examples{
print(meses)
}
\keyword{expresiones}
\keyword{regulares}
