% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notas.R
\docType{data}
\name{notas}
\alias{notas}
\title{Colección de notas de la Revista Puerto.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7816 rows and 6 columns.
}
\usage{
notas
}
\description{
Función que carga más de 8000 notas sobre cuestiones pesqueras para el período 2009-2021.
}
\examples{
notas



}
\keyword{datasets}
