% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_puerto.R
\docType{data}
\name{rev_puerto}
\alias{rev_puerto}
\title{Colección de notas}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7816 rows and 6 columns.
}
\usage{
rev_puerto
}
\description{
Colección de notas de la Revista Puerto de Mar del Plata, Argentina. Es una base de datos en formato tabular con 7816 observaciones y 6 variables: fecha, titulo, bajada, nota, imagen, link.
}
\examples{
notas_rev_puerto <- rev_puerto
head_rev_puerto <- head(rev_puerto, 50)
}
\references{
Nieto, Agustín. 2020. «Intersecciones entre historia digital e historia social: un ejercicio de lectura distante sobre la conflictividad marítima en la historia argentina reciente». Drassana: revista del Museu Marítim (28):122-42.
}
\keyword{datasets}
