% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sintildes.R
\docType{data}
\name{sintildes}
\alias{sintildes}
\title{Listado de letras sin tildes}
\format{
Es una cadena de caracteres.
\describe{
\item{sintildes}{es un string}
}
}
\usage{
data(sintildes)
}
\description{
Lista de letras sin tildes.
}
\examples{
print(sintildes)
}
\keyword{expresiones}
\keyword{regulares}
