% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{Filtro para remover stopwords}
\format{
Es una cadena de caracteres.
\describe{
\item{stopwords}{es un string}
}
}
\usage{
data(stopwords)
}
\description{
Lista de palabras para remover stopwords.
}
\examples{
print(stopwords)
}
\keyword{expresiones}
\keyword{regulares}
