% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tildes.R
\docType{data}
\name{tildes}
\alias{tildes}
\title{Filtro para remover tildes}
\format{
Es una cadena de caracteres.
\describe{
\item{tildes}{es un string}
}
}
\usage{
data(tildes)
}
\description{
Lista de palabras para remover tildes.
}
\examples{
print(tildes)
}
\keyword{expresiones}
\keyword{regulares}
