% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_bases.R
\docType{data}
\name{acep_bases}
\alias{acep_bases}
\title{Coleccion de notas}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7816 rows and 6 columns.
}
\usage{
acep_bases
}
\description{
Coleccion de notas de la Revista Puerto de Mar del Plata, Argentina. Es una base de datos en formato tabular con 7816 observaciones y 6 variables: fecha, titulo, bajada, nota, imagen, link.
}
\examples{
notas_rev_puerto <- acep_bases
head_rev_puerto <- head(acep_bases, 50)
}
\references{
Nieto, Agustin 2020 «Intersecciones entre historia digital e historia social: un ejercicio de lectura distante sobre la conflictividad maritima en la historia argentina reciente». Drassana: revista del Museu Maritim (28):122-42.
}
\keyword{datasets}
