% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_clean.R
\name{acep_clean}
\alias{acep_clean}
\title{Limpieza de texto.}
\usage{
acep_clean(
  x,
  tolower = TRUE,
  rm_cesp = TRUE,
  rm_emoji = TRUE,
  rm_hashtag = TRUE,
  rm_users = TRUE,
  rm_punt = TRUE,
  rm_num = TRUE,
  rm_url = TRUE,
  rm_meses = TRUE,
  rm_dias = TRUE,
  rm_stopwords = TRUE,
  rm_shortwords = TRUE,
  rm_newline = TRUE,
  rm_whitespace = TRUE,
  other_sw = NULL,
  u = 1
)
}
\arguments{
\item{x}{vector de textos al que se le aplica la
funcion de limpieza de texto.}

\item{tolower}{convierte los textos a minusculas.}

\item{rm_cesp}{remueve caracteres especiales.}

\item{rm_emoji}{remueve los emojis.}

\item{rm_hashtag}{remueve los hashtags.}

\item{rm_users}{remueve las menciones de usuarixs de redes sociales.}

\item{rm_punt}{remueve la puntuacion.}

\item{rm_num}{remueve numeros.}

\item{rm_url}{remueve las url.}

\item{rm_meses}{remueve los meses del anio.}

\item{rm_dias}{remueve los dias de la semana.}

\item{rm_stopwords}{remueve palabras vacias.}

\item{rm_shortwords}{remueve las palabras cortas.}

\item{rm_newline}{remueve los saltos de linea.}

\item{rm_whitespace}{remueve los espacios en blanco.}

\item{other_sw}{su valor por defecto es NULL, sirve para ampliar el
listado de stopwords con un nuevo vector de palabras.}

\item{u}{umbral de caracteres para la funcion rm_shortwords.}
}
\description{
Funcion que limpia y normaliza las notas/textos.
}
\examples{
acep_clean("El SUTEBA fue al paro. Reclaman mejoras salariales.", rm_cesp = FALSE)
}
\keyword{normalizacion}
