% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_context.R
\name{acep_context}
\alias{acep_context}
\title{Funcion para extraer contexto de palabras o frases.}
\usage{
acep_context(texto, clave, izq = 1, der = 1)
}
\arguments{
\item{texto}{vector con los textos a procesar.}

\item{clave}{vector de palabras clave a contextualizar.}

\item{izq}{numero de palabras de la ventana hacia la izquierda.}

\item{der}{numero de palabras de la ventana hacia la derecha.}
}
\value{
Si todas las entradas son correctas,
la salida sera un data frame con el id de los textos
procesado y el contexto de las palabras y/o frases entradas.
}
\description{
Funcion que devuelve un data.frame con el contexto
de una o más palabras o frases según una ventana de palabras hacia
las izquierda y otra ventana de palabras hacia la derecha.
}
\examples{
texto <- "El SOIP para por aumento de salarios"
texto_context <- acep_context(texto = texto, clave = "para")
texto_context
}
\keyword{contexto}
