% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_context.R
\name{acep_context}
\alias{acep_context}
\title{Función para extraer contexto de palabras o frases.}
\usage{
acep_context(texto, clave, izq = 1, der = 1, ci = "\\\\b", cd = "\\\\S*")
}
\arguments{
\item{texto}{vector con los textos a procesar.}

\item{clave}{vector de palabras clave a contextualizar.}

\item{izq}{número de palabras de la ventana hacia la izquierda.}

\item{der}{número de palabras de la ventana hacia la derecha.}

\item{ci}{expresión regular a la izquierda de la palabra clave.}

\item{cd}{expresión regular a la derecha de la palabra clave.}
}
\value{
Si todas las entradas son correctas,
la salida sera un data frame con el id de los textos
procesado y el contexto de las palabras y/o frases entradas.
}
\description{
Función que devuelve un data.frame con el contexto
de una o más palabras o frases según una ventana de palabras hacia
las izquierda y otra ventana de palabras hacia la derecha.
}
\examples{
texto <- "El SOIP para por aumento de salarios"
texto_context <- acep_context(texto = texto, clave = "para")
texto_context
}
\keyword{contexto}
\keyword{tokens}
