% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_db.R
\name{acep_db}
\alias{acep_db}
\title{Frecuencia, menciones e intensidad}
\usage{
acep_db(db, t, d, n)
}
\arguments{
\item{db}{data frame con los textos a procesar.}

\item{t}{columna de data frame que contiene el vector de textos a procesar.}

\item{d}{diccionario en formato vector.}

\item{n}{cantidad de decimales del índice de intensidad.}
}
\description{
Función que usa las funciones acep_frec, acep_men y acep_int y devuelve una tabla con tres columnas nuevas: número de palabras, número de menciones del diccionario, índice de intensidad.
}
\examples{
rp_procesada <- acep_db(acep_bases$rev_puerto, acep_bases$rev_puerto$nota, acep_diccionarios$dicc_viol_gp, 4)
}
