% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_db.R
\name{acep_db}
\alias{acep_db}
\title{Frecuencia, menciones e intensidad.}
\usage{
acep_db(db, t, d, n)
}
\arguments{
\item{db}{data frame con los textos a procesar.}

\item{t}{columna de data frame que contiene el vector
de textos a procesar.}

\item{d}{diccionario en formato vector.}

\item{n}{cantidad de decimales del indice de intensidad.}
}
\value{
Si todas las entradas son correctas, la salida sera una
base de datos en formato tabular con tres nuevas variables.
}
\description{
Función que usa las funciones acep_frec, acep_count y acep_int y
devuelve una tabla con tres columnas nuevas: numero de palabras,
número de menciones del diccionario, indice de intensidad.
}
\examples{
df <- data.frame(texto = c("El SUTEBA fue al paro. Reclaman mejoras salariales.",
"El SOIP lleva adelante un plan de lucha con paros y piquetes."))
diccionario <- c("paro", "lucha", "piquetes")
acep_db(df, df$texto, diccionario, 4)
}
\keyword{frecuencia}
\keyword{indicadores}
\keyword{tokens}
