% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_extract.R
\name{acep_extract}
\alias{acep_extract}
\title{Función para buscar y extraer palabras en un texto.}
\usage{
acep_extract(texto, dic, sep = "; ", izq = "\\\\b\\\\w*", der = "\\\\w*\\\\b")
}
\arguments{
\item{texto}{El texto en el que se buscaran las palabras clave.}

\item{dic}{Un vector con las palabras clave a buscar.}

\item{sep}{El separador utilizado para concatenar las
palabras clave encontradas (por defecto: " | ").}

\item{izq}{expresión regular para incorporar otros caracteres a
la izquierda de los términos del vector 'dic'.}

\item{der}{expresión regular para incorporar otros caracteres a
la derecha de los términos del vector 'dic'.}
}
\value{
Si todas las entradas son correctas,
la salida sera un vector de tipo caracter.
procesado y el contexto de las palabras y/o frases entradas.
}
\description{
Esta función busca palabras clave en un texto y
extrae los resultados en un formato especifico.
}
\examples{
texto <- "Los obreros del pescado, en el marco de una huelga y
realizaron una manifestación con piquete en el puerto de la ciudad."
dicc <- c("huel", "manif", "piq")
acep_extract(texto, dicc)
}
\keyword{buscar}
\keyword{diccionario}
\keyword{palabras}
\keyword{texto}
