% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_frec.R
\name{acep_frec}
\alias{acep_frec}
\title{Frecuencia de palabras totales.}
\usage{
acep_frec(x)
}
\arguments{
\item{x}{vector de textos al que se le aplica la funcion de conteo
de la frecuencia de palabras.}
}
\value{
Si todas las entradas son correctas, la salida sera un vector
con una frecuencia de palabras.
}
\description{
Funcion que cuenta la frecuencia de palabras totales
en cada una de las notas/textos.
}
\examples{
acep_frec("El SUTEBA fue al paro. Reclaman mejoras salariales.")
}
\keyword{indicadores}
