% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_frec.R
\name{acep_frec}
\alias{acep_frec}
\title{Frecuencia de menciones de palabras.}
\usage{
acep_frec(x, y, tolower = TRUE)
}
\arguments{
\item{x}{vector de textos al que se le aplica la función de conteo de la frecuencia de menciones de palabras del diccionario.}

\item{y}{vector de palabras del diccionario utilizado.}

\item{tolower}{convierte los textos a minúsculas.}
}
\description{
Función que cuenta la frecuencia de menciones de palabras que refieren a conflictos en cada una de las notas/textos.
}
\examples{
rev_puerto$conflictos <- acep_frec(rev_puerto$nota, diccionarios$dicc_violencia)

}
