% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_frec.R
\name{acep_frec}
\alias{acep_frec}
\title{Frecuencia de palabras totales.}
\usage{
acep_frec(x)
}
\arguments{
\item{x}{vector de textos al que se le aplica la función de conteo de la frecuencia de palabras.}
}
\value{
Si todas las entradas son correctas, la salida será un vector con una frecuencia de palabras.
}
\description{
Función que cuenta la frecuencia de palabras totales en cada una de las notas/textos.
}
\examples{
rev_puerto <- rev_puerto <- acep_bases$rev_puerto
rev_puerto$n_palabras <- acep_frec(rev_puerto$nota)
rev_puerto |> head()
}
\keyword{indicadores}
