% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_gpt.R
\name{acep_gpt}
\alias{acep_gpt}
\title{Función para interactuar con los modelos  de OpenAI.}
\usage{
acep_gpt(
  texto,
  instrucciones,
  gpt_api = Sys.getenv("OPENAI_API_KEY"),
  modelo = "gpt-3.5-turbo-0125"
)
}
\arguments{
\item{texto}{texto a procesar.}

\item{instrucciones}{indicaciones de contexto que se le
proporcionan al modelo de lenguaje GPT para influir en
su generación de texto.}

\item{gpt_api}{clave API secreta.}

\item{modelo}{modelo de lenguaje GPT de OpenAI.}
}
\value{
Si todas las entradas son correctas,
la salida sera un JSON con la información solicitada.
}
\description{
Función para interactuar con los modelos de OpenAI
desde las claves API secretas y pagas.
}
\examples{
\dontrun{
texto <- "El SOIP declaro la huelga por aumento de salarios."
pregunta <- "¿Cuál es el sujeto de la acción? Ejemplo de respuesta: 'CGT'"
texto_gpt <- acep_gpt(texto = texto, instrucciones = pregunta)
cat(texto_gpt)
}
}
\keyword{gpt}
\keyword{tripletes}
