% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_int.R
\name{acep_int}
\alias{acep_int}
\title{Índice de intensidad.}
\usage{
acep_int(pc, pt, decimales = 4)
}
\arguments{
\item{pc}{vector numérico con la frecuencia de palabras conflictivas presentes en cada texto.}

\item{pt}{vector de palabras totales en cada texto.}

\item{decimales}{cantidad de decimales, por defecto tiene 4 pero se puede modificar.}
}
\description{
Función que elabora un indice de intensidad en base a la relación entre palabras totales y palabras del diccionario presentes en el texto.
}
\examples{
acep_bases$rev_puerto$intensidad <- acep_int(acep_bases$rev_puerto$conflictos, acep_bases$rev_puerto$n_palabras, 3)
}
