% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_int.R
\name{acep_int}
\alias{acep_int}
\title{Indice de intensidad.}
\usage{
acep_int(pc, pt, decimales = 4)
}
\arguments{
\item{pc}{vector numerico con la frecuencia de palabras conflictivas
presentes en cada texto.}

\item{pt}{vector de palabras totales en cada texto.}

\item{decimales}{cantidad de decimales, por defecto tiene 4
pero se puede modificar.}
}
\value{
Si todas las entradas son correctas,
la salida sera un vector numerico.
}
\description{
Funcion que elabora un indice de intensidad en
base a la relacion entre palabras totales y palabras del diccionario
presentes en el texto.
}
\examples{
conflictos <- c(1, 5, 0, 3, 7)
palabras <- c(4, 11, 12, 9, 34)
acep_int(conflictos, palabras, 3)
}
\keyword{indicadores}
