% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_load_base.R
\name{acep_load_base}
\alias{acep_load_base}
\title{Carga bases de datos creadas por el Observatorio.}
\usage{
acep_load_base(tag)
}
\arguments{
\item{tag}{etiqueta identificatoria del data frame a cargar: 'rp_mdp', 'ln_bb'}
}
\description{
Función para cargar bases de datos disponibles online. Por ahora están disponibles las siguientes bases: Revista Puerto 'rp_mdp'; La Nueva 'ln_bb'
}
\examples{
\dontrun{acep_load_base('rp_mdp') |> head()}
}
\keyword{datos}
