% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_load_base.R
\name{acep_load_base}
\alias{acep_load_base}
\title{Carga bases de datos creadas por el Observatorio.}
\usage{
acep_load_base(tag)
}
\arguments{
\item{tag}{etiqueta identificatoria del data frame a cargar: acep_bases$rp_mdp, acep_bases$ln_bb, acep_bases$lc_mdp, acep_bases$ed_neco, acep_bases$ln_arg}
}
\value{
Si todas las entradas son correctas, la salida ser치 una base de datos en formato tabular con un corpus de notas.
}
\description{
Funci칩n para cargar bases de datos disponibles online. Por ahora est치n disponibles las siguientes bases: Revista Puerto 'rp_mdp'; La Nueva 'ln_bb', La Capital 'lc_mdp', Ecos Diarios 'ed_neco', La Naci칩n 'ln_arg'
}
\examples{
bd_sismos <- 'https://zenodo.org/record/6835713/files/bd_sismos_mdp.rds?download=1'
acep_load_base(tag = bd_sismos) |> head()
}
\keyword{datos}
