% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_plot_rst.R
\name{acep_plot_rst}
\alias{acep_plot_rst}
\title{Resumen visual de la serie temporal de los indices de conflictividad.}
\usage{
acep_plot_rst(datos, tagx = "horizontal")
}
\arguments{
\item{datos}{data frame con datos procesados.}

\item{tagx}{orientaci칩n de las etiquetas del
eje x ('horizontal' | 'vertical').}
}
\value{
Si todas las entradas son correctas,
la salida sera una imagen de cuatro paneles.
}
\description{
Funci칩n que devuelve un panel visual de cuatro gr치ficos
de barras con variables proxy de los indices de conflictividad agrupados
por segmento de tiempo.
}
\examples{
datos <- acep_bases$rp_procesada
fecha <- datos$fecha
n_palabras <- datos$n_palabras
conflictos <- datos$conflictos
datos_procesados_anio <- acep_rst(datos,
fecha, n_palabras, conflictos, st = 'anio')
acep_plot_rst(datos_procesados_anio, tagx = 'vertical')
}
\keyword{visualizaci칩n}
