% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_plot_st.R
\name{acep_plot_st}
\alias{acep_plot_st}
\title{Serie temporal de índices de conflictividad.}
\usage{
acep_plot_st(x, y, t = "", ejex = "", ejey = "", etiquetax = "horizontal")
}
\arguments{
\item{x}{vector de valores del eje x (por ejemplo, fechas).}

\item{y}{vector de valores numéricos del eje y (por ejemplo, menciones).}

\item{t}{titulo del gráfico.}

\item{ejex}{nombre del eje x.}

\item{ejey}{nombre del eje y.}

\item{etiquetax}{orientación de las etiquetas del eje x ('horizontal' | 'vertical').}
}
\description{
Función que devuelve los índices de conflictividad agrupados por segmento de tiempo: día, mes, año.
}
\examples{
datos <- acep_db(acep_bases$rev_puerto, acep_bases$rev_puerto$nota, acep_diccionarios$dicc_viol_gp, 4)
datos_procesados_anio <- acep_rst(datos, datos$fecha, datos$n_palabras, datos$conflictos, st = 'anio')
acep_plot_st(datos_procesados_anio$st, datos_procesados_anio$frecm,
             t = 'Evolución de la conflictividad en el sector pesquero argentino',
             ejex = 'Años analizados',
             ejey = 'Menciones de términos del diccionario de conflictos',
             etiquetax = 'horizontal')
}
