% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_plot_st.R
\name{acep_plot_st}
\alias{acep_plot_st}
\title{Gráfico de barras de la serie temporal de indices de conflictividad.}
\usage{
acep_plot_st(
  x,
  y,
  t = "",
  ejex = "",
  ejey = "",
  etiquetax = "horizontal",
  color = "mint"
)
}
\arguments{
\item{x}{vector de valores del eje x (por ejemplo, fechas).}

\item{y}{vector de valores numéricos del eje y (por ejemplo, menciones).}

\item{t}{titulo del gráfico.}

\item{ejex}{nombre del eje x.}

\item{ejey}{nombre del eje y.}

\item{etiquetax}{orientación de las etiquetas del
eje x ('horizontal' | 'vertical').}

\item{color}{color de las barras.}
}
\value{
Si todas las entradas son correctas,
la salida sera una imagen de un panel.
}
\description{
Función que devuelve un gráfico de barras con la serie
temporal de indices de conflictividad por dia, mes o anio.
}
\examples{
datos <- acep_bases$rp_procesada
dpa <- acep_sst(datos, st = 'anio')
acep_plot_st(
dpa$st, dpa$frecm,
t = 'Evoluci\u00f3n de la conflictividad en el sector pesquero argentino',
ejex = 'A\u00f1os analizados',
ejey = 'Menciones de t\u00e9rminos del diccionario de conflictos',
etiquetax = 'horizontal')
}
\keyword{visualizacion}
