% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_postag.R
\name{acep_postag}
\alias{acep_postag}
\title{Función para etiquetado POS, lematización, tokenización, extracción de entidades.}
\source{
\href{https://universaldependencies.org/}{Dependencias Universales para taggeo POS}

\href{https://cran.r-project.org/web/packages/spacyr/spacyr.pdf}{Sobre el modelo spaCy de Python para R}

\href{https://jessecambon.github.io/tidygeocoder/articles/tidygeocoder.html}{Sobre el paquete tidygeocoder}
}
\usage{
acep_postag(
  texto,
  core = "es_core_news_lg",
  bajar_core = TRUE,
  inst_spacy = FALSE,
  inst_miniconda = FALSE,
  inst_reticulate = FALSE
)
}
\arguments{
\item{texto}{vector con los textos a procesar.}

\item{core}{idioma del modelo de etiquetado POS del paquete {spacyr}.}

\item{bajar_core}{parámetro booleano que define si bajar o no el modelo de etiquetado POS.}

\item{inst_spacy}{parámetro booleano que define si instalar o no spacy (python).}

\item{inst_miniconda}{parámetro booleano que define si instalar o no miniconda.
Miniconda es un instalador mínimo gratuito para conda. Es una pequeña versión de
arranque de Anaconda que incluye solo conda, Python, los paquetes de los que dependen y
una pequeña cantidad de otros paquetes útiles, incluidos pip, zlib y algunos otros.}

\item{inst_reticulate}{parámetro booleano que define si instalar o no el paquete {reticulate}.}
}
\value{
Si todas las entradas son correctas, la salida sera una lista con seis bases de datos en formato tabular.
}
\description{
Función que devuelve seis objetos data.frame con
etiquetado POS (modelo spacyr) para su posterior
procesamiento con la función acep_postag.
}
\examples{
\dontrun{
texto <- "En Mar del Plata el SOIP declara la huelga en demanda de aumento salarial."
texto_postag <- acep_postag(texto)
texto_postag$texto_tag[ , c(1:8)]
}
}
\keyword{etiquetado}
