% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_prompt_gpt.R
\docType{data}
\name{acep_prompt_gpt}
\alias{acep_prompt_gpt}
\title{Colección de instrucciones para GPT.}
\format{
Es un objeto de clase 'list' con 1 componente.
\describe{
\item{dicc_confl_gp}{es un texto con instrucciones para que el modelo
 gpt-3 turbo de OpenAI devuelva una codificación de eventos de protesta.}
}
}
\usage{
data(acep_prompt_gpt)
}
\description{
Colección de instrucciones para interactuar con los modelos
de OpenAI. Las instrucciones fueron testeadas en el marco de las tareas
que realizamos en el Observatorio de Conflictividad Social de la
Universidad Nacional de Mar del Plata.
}
\examples{
prompt01 <- acep_prompt_gpt$instrucciones_gpt
prompt01
}
\keyword{diccionarios}
