% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_rs.R
\docType{data}
\name{acep_rs}
\alias{acep_rs}
\title{Cadenas de caracteres para limpiar y normalizar textos.}
\format{
Son cadenas de caracteres.
\describe{
\item{stopwords}{es un string de palabras vacias.}
\item{dias}{es un string de dias.}
\item{meses}{es un string de meses.}
\item{emoji}{es un string con expresiones regulares para emojis.}
\item{sintildes}{es un string de letras sin tildes.}
\item{tildes}{es un string de letras con tildes.}
}
}
\usage{
data(acep_rs)
}
\description{
Cadenas de caracteres y expresiones regulares
para limpiar y normalizar textos.
}
\examples{
print(acep_rs)
}
\keyword{expresiones}
\keyword{regulares}
