% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_svo.R
\name{acep_svo}
\alias{acep_svo}
\title{Funcion para extraer triplets SVO (Sujeto-Verbo-Objeto).}
\usage{
acep_svo(
  texto,
  prof_s = 1,
  prof_o = 1,
  rel_s = c("nsubj", "conj"),
  rel_o = c("obj", "obl", "nmod")
)
}
\arguments{
\item{texto}{vector con los textos a procesar.}

\item{prof_s}{profundidad de relaciones sintacticas para el sujeto.}

\item{prof_o}{profundidad de relaciones sintacticas para el objeto.}

\item{rel_s}{vector de etiquetas de relaciones sintacticas en el sujeto.}

\item{rel_o}{vector de etiquetas de relaciones sintacticas en el objeto.}
}
\value{
Si todas las entradas son correctas,
la salida sera una lista con tres bases de datos en formato tabular.
}
\description{
Funcion que devuelve tres objetos data.frame con
etiquetado POS y relaciones sintacticas que permiten reconstruir
estructuras sintacticas como SVO y Sujeto-Predicado
}
\examples{
texto <- "El SOIP declara la huelga en demanda de aumento salarial."
texto_svo <- acep_svo(texto)
texto_svo
}
\keyword{sintaxis}
