% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_svo.R
\name{acep_svo}
\alias{acep_svo}
\title{Funcion para extraer triplets SVO (Sujeto-Verbo-Objeto).}
\usage{
acep_svo(
  texto,
  modelo = "spanish",
  prof_s = 2,
  prof_o = 2,
  relaciones = c("flat", "fixed", "compount"),
  conexiones = FALSE,
  rel_s = c("nsubj", "conj"),
  rel_o = c("obj", "obl", "nmod"),
  rel_evs = "nsubj",
  rel_evp = "obj",
  u = 1
)
}
\arguments{
\item{texto}{vector con los textos a procesar.}

\item{modelo}{idioma del modelo de etiquetado POS del paquete {udpipe}.}

\item{prof_s}{profundidad de relaciones sintacticas para el sujeto.}

\item{prof_o}{profundidad de relaciones sintacticas para el objeto.}

\item{relaciones}{vector con las relaciones de analisis sintactico del paquete {rsyntax}.}

\item{conexiones}{es un parámetro heredado del paquete {rsyntax} que controla
el comportamiento si la profundidad es > 1 y se utilizan filtros.
Si es FALSE, se recuperan todos los padres/hijos hasta la profundidad dada, y luego se filtran.
De esta manera, los nietos que satisfacen las condiciones del filtro se recuperan incluso si
sus padres no satisfacen las condiciones. Si es TRUE, el filtro se aplica en cada nivel de
profundidad, de modo que sólo se recuperan las ramas totalmente conectadas de los nodos que
satisfacen las condiciones.}

\item{rel_s}{vector de etiquetas de relaciones sintacticas en el sujeto.}

\item{rel_o}{vector de etiquetas de relaciones sintacticas en el objeto.}

\item{rel_evs}{etiqueta de relaciones a ser agregada en la reconstruccion del sujeto.}

\item{rel_evp}{etiqueta de relaciones a ser agregada en la reconstruccion del objeto.}

\item{u}{numero entero que indica el umbral de palabras del objeto en la reconstruccion SVO.}
}
\value{
Si todas las entradas son correctas,
la salida sera una lista con tres bases de datos en formato tabular.
}
\description{
Funcion que devuelve cinco objetos data.frame con
etiquetado POS (modelo udpipe) y relaciones sintacticas
(modelo rsyntax) que permiten reconstruir estructuras
sintacticas como SVO y Sujeto-Predicado. Una vez seleccionadas
las notas periodisticas referidas a conflictos, esta función permite
extraer sujetos de la protesta, acción realizada y objeto(s) de la acción.
}
\examples{
texto <- "El SOIP declara la huelga en demanda de aumento salarial."
texto_svo <- acep_svo(texto)
texto_svo$acep_list_svo
}
\keyword{sintaxis}
