% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_token.R
\name{acep_token}
\alias{acep_token}
\title{Tokenizador.}
\usage{
acep_token(x, tolower = TRUE)
}
\arguments{
\item{x}{vector de textos al que se le aplica la funcion de tokenizacion.}

\item{tolower}{convierte los textos a minusculas.}
}
\description{
Funcion que tokeniza las notas/textos.
}
\examples{
acep_token("Huelga de obreros del pescado en el puerto")
}
\keyword{tokenizar}
