% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_token.R
\name{acep_token}
\alias{acep_token}
\title{Tokenizador.}
\usage{
acep_token(x, tolower = TRUE)
}
\arguments{
\item{x}{vector de textos al que se le aplica la función de tokenización.}

\item{tolower}{convierte los textos a minúsculas.}
}
\description{
Función que tokeniza las notas/textos.
}
\examples{
rev_puerto <- acep_bases$rev_puerto
rev_puerto_token <- acep_token(rev_puerto$titulo)
rev_puerto_token |> head()
}
\keyword{tokenizar}
