% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_token_plot.R
\name{acep_token_plot}
\alias{acep_token_plot}
\title{Gráfico de barras de palabras más recurrentes en un corpus.}
\usage{
acep_token_plot(x, u = 10, frec = TRUE)
}
\arguments{
\item{x}{vector de palabras tokenizadas.}

\item{u}{numero de corte para el top de palabras más frecuentes.}

\item{frec}{parámetro para determinar si los valores se visualizarán
como frecuencia absoluta o relativa.}
}
\value{
Si todas las entradas son correctas,
la salida será un gráfico de barras.
}
\description{
Función que devuelve un gráfico de barras con las palabras
más recurrentes en un corpus textual.
}
\examples{
rev_puerto <- acep_clean(acep_bases$rev_puerto$titulo)
rev_puerto_token <- acep_token(rev_puerto)
acep_token_plot(rev_puerto_token$token)
}
\keyword{visualización}
