% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_token_table.R
\name{acep_token_table}
\alias{acep_token_table}
\title{Tabla de frecuencia de palabras tokenizadas.}
\usage{
acep_token_table(x, u = 10)
}
\arguments{
\item{x}{vector de palabras tokenizadas.}

\item{u}{numero de corte para el top de palabras más frecuentes.}
}
\value{
Si todas las entradas son correctas,
la salida será una tabla con la frecuencia relativa y
absoluta de palabras tokenizadas.
}
\description{
Función que cuenta la frecuencia de palabras tokenizadas.
}
\examples{
rev_puerto <- acep_clean(acep_bases$rev_puerto$titulo)
rev_puerto_token <- acep_token(rev_puerto)
acep_token_table(rev_puerto_token$token)
}
\keyword{tablas}
