% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acep_upos.R
\name{acep_upos}
\alias{acep_upos}
\title{Función para etiquetado POS, lematización, tokenización.}
\source{
\href{https://universaldependencies.org/}{Dependencias Universales para taggeo POS}

\href{https://ufal.mff.cuni.cz/~straka/papers/2017-conll_udpipe.pdf}{Sobre el modelo UDPipe}

\href{https://cran.r-project.org/web/packages/rsyntax/rsyntax.pdf}{Sobre el paquete rsyntax}
}
\usage{
acep_upos(texto, modelo = "spanish")
}
\arguments{
\item{texto}{vector con los textos a procesar.}

\item{modelo}{idioma del modelo de etiquetado POS del paquete \code{udpipe}.}
}
\value{
Si todas las entradas son correctas, la salida sera un marco de datos con 17 variables.
}
\description{
Función que devuelve un marco de datos objetos con
etiquetado POS (modelo udpipe) para su posterior
procesamiento con la función acep_postag.
}
\examples{
\dontrun{
texto <- "El SOIP declara la huelga en demanda de aumento salarial."
acep_upos(texto)
}
}
\references{
Welbers, K., Atteveldt, W. van, & Kleinnijenhuis, J. 2021. Extracting semantic relations using syntax: An R package for querying and reshaping dependency trees. Computational Communication Research, 3-2, 1-16.
\href{https://www.aup-online.com/content/journals/10.5117/CCR2021.2.003.WELB?TRACK}{(link al articulo)}
}
\keyword{sintaxis}
