% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionarios.R
\docType{data}
\name{diccionarios}
\alias{diccionarios}
\title{Colección de diccionarios.}
\format{
An object of class \code{list} of length 2.
}
\usage{
diccionarios
}
\description{
Función que carga un set de diccionarios de diferentes orígenes.
Los diccionarios de palabras que refieren a conflictos y a violencia llamados
dicc_conflictos y dicc_violencia fueron extraídos de Albrieu y Palazzo (2020).
}
\examples{
dicc_conflictos <- diccionarios$dicc_conflictos
dicc_violencia <-diccionarios$dicc_violencia
}
\references{
Albrieu, Ramiro, y Gabriel Palazzo. 2020. «Categorización de conflictos sociales en el ámbito de los recursos naturales: un estudio de las actividades extractivas mediante la minería de textos». Revista CEPAL (131):29-59.
}
\keyword{datasets}
